% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{Find the difference (delta)}
\usage{
delta(data, x, y)
}
\arguments{
\item{data}{is a data set that was filtered by observations
and have each column for maximum and minimum values.}

\item{x}{is the column of maximum values.}

\item{y}{is the column of minimum values.}
}
\value{
Returns a table of outputs that are the difference between maximum and minimum values for each observation.
}
\description{
This function is to create a table of delta values, which is the difference
between the annual maximum value and minimum value.
}
\examples{
#I will create a data set that has maximum and minimum values for each year.

library(tidyverse)
library(datateachr)
pivoted_flow_sample <- flow_sample \%>\%
pivot_wider("year", names_from = "extreme_type", values_from = "flow") \%>\%
group_by(year) \%>\%
mutate(maximum = as.numeric(maximum)) \%>\%
mutate(minimum = as.numeric(minimum))

#Using the created data set, "pivoted_flow_sample", I will create a table of outputs,
#that are annual differences between the maximum and minimum values.

delta(pivoted_flow_sample, "maximum", "minimum")

}
